Move logger function outside of utils
